requirements.txt
numpy
newspaper3k
nltk
scikit_learn
